# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RRHH API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/puesto-subfamilias": {
            "get": {
                "tags": [
                    "Puesto Subfamilias"
                ],
                "summary": "getPuestoSubfamiliaList",
                "description": "Get all PuestoSubfamilias",
                "operationId": "dd2eed6cb9a0e89b1a4f93b263cc69a6",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Puesto Subfamilias"
                ],
                "summary": "createPuestoSubfamilia",
                "description": "Create PuestoSubfamilia",
                "operationId": "5ef80e608d44bf403f71e98bf04f1f8b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    ""
                                ],
                                "properties": {
                                    "name": {
                                        "description": "desc",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/puesto-subfamilias/{id}": {
            "get": {
                "tags": [
                    "Puesto Subfamilias"
                ],
                "summary": "getPuestoSubfamiliaItem",
                "description": "Get PuestoSubfamilia",
                "operationId": "7172ebab90aa32b396205a79613a1365",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of PuestoSubfamilia",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Puesto Subfamilias"
                ],
                "summary": "updatePuestoSubfamilia",
                "description": "Update PuestoSubfamilia",
                "operationId": "8e8d4e72c59dd239064baca2fb7786c6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of PuestoSubfamilia",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    ""
                                ],
                                "properties": {
                                    "name": {
                                        "description": "desc",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Puesto Subfamilias"
                ],
                "summary": "deletePuestoSubfamilia",
                "description": "Delete PuestoSubfamilia",
                "operationId": "b67db691e32cff35c5dd81fca88224fb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of PuestoSubfamilia",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "in": "header"
            }
        }
    }
}